+++ frontend/src/components/common/export-recipients/ExportRecipients.module.scss
@@ -31,3 +31,25 @@
 .disabled {
   opacity: 0.5;
 }
+
+.actionButton {
+  @include flex-horizontal;
+  justify-content: flex-end;
+  margin: 3rem 0;
+
+  .disableActiveState {
+    cursor: not-allowed;
+
+    &:active {
+      span,
+      i {
+        color: $indigo-blue;
+      }
+    }
+
+    &:hover {
+      border-width: 1px;
+      background-color: white;
+    }
+  }
+}
