+++ frontend/src/components/common/progress-details/ProgressDetails.module.scss
@@ -79,30 +79,3 @@
     font-size: 1rem;
   }
 }
-
-.statsLastUpdated {
-  @include flex-horizontal;
-  justify-content: space-between;
-}
-
-.actionButton {
-  @include flex-horizontal;
-  justify-content: flex-end;
-  margin: 3rem 0;
-
-  .disableActiveState {
-    cursor: not-allowed;
-
-    &:active {
-      span,
-      i {
-        color: $indigo-blue;
-      }
-    }
-
-    &:hover {
-      border-width: 1px;
-      background-color: white;
-    }
-  }
-}
