+++ frontend/src/components/dashboard/campaigns/Campaigns.tsx
@@ -13,7 +13,7 @@ import {
   PrimaryButton,
   ExportRecipients,
 } from 'components/common'
-import { getCampaigns, getExportStatus } from 'services/campaign.service'
+import { getCampaigns } from 'services/campaign.service'
 import { Campaign, channelIcons, Status } from 'classes'
 import CreateCampaign from 'components/dashboard/create/create-modal'
 
@@ -114,11 +114,8 @@ const Campaigns = () => {
             campaignName={campaign.name}
             status={campaign.status}
             sentAt={campaign.sentAt}
-            exportStatus={getExportStatus(
-              campaign.status,
-              campaign.statusUpdatedAt,
-              +campaign.hasFailedRecipients
-            )}
+            updatedAt={campaign.statusUpdatedAt}
+            hasFailedRecipients={campaign.hasFailedRecipients}
           />
         )
       },
